Add basic flake and nix build providing source files
Adds a very basic flake.nix file as well with a basic nix build
script in default.nix that just provides the Zig source files.

This makes it possible for other Nix flake projects to import this
package and build it as part of their nix zig project by accessing
the zig source.